% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_fig}
\alias{save_fig}
\title{Save figure}
\usage{
save_fig(p, file_name, width = NULL, height = NULL)
}
\arguments{
\item{p}{plot}

\item{file_name}{file name}

\item{width}{width}

\item{height}{height}
}
\description{
Save figure
}
